+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes in version 2.1?.? - 2020-0?-??
 
+ * Minor changes
+   - Avoid invoking overridable methods from constructors.
+
 
 # Changes in version 2.10.0 - 2020-01-15